// src/types/boleto-type.ts
export type BoletoType =
  | "CARTÃO_DE_CRÉDITO"
  | "ARRECADAÇÃO_PREFEITURA"
  | "CONVÊNIO_SANEAMENTO"
  | "CONVÊNIO_ENERGIA_ELÉTRICA_E_GAS"
  | "CONVÊNIO_TELECOMUNICAÇÕES"
  | "ARRECADAÇÃO_ÓRGÃOS_GOVERNAMENTAIS"
  | "OUTROS"
  | "ARRECADAÇÃO_TAXAS_DE_TRANSITO"
  | "OUTROS"
  | "BANCO"